+++ .idea/modules.xml
@@ -2,6 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/CO600-Project.iml" filepath="$PROJECT_DIR$/CO600-Project.iml" />
       <module fileurl="file://$PROJECT_DIR$/KentApp.iml" filepath="$PROJECT_DIR$/KentApp.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
