--- app/src/main/java/kent/kentapp/DataArrays.java
@@ -1,11 +1,0 @@
-package kent.kentapp;
-
-public class DataArrays {
-    static String[] medicalCentre = {"University Medical Centre" + "Giles Lane\n" +
-            "Canterbury\n" + "Kent\n" + "CT2 7PB\n" + "Telephone: 01227 469333\n" + "http://www.umckent.co.uk/\n"};
-    static String[] finances = {"Finance Department" + "The Registry\n" + "University of Kent\n" + "Canterbury\n" +
-            "Kent\n" + "Fax: +44(0)1227 827024\n"};
-    static String[] security = {"Canterbury Campus" + "University of Kent,\n" +
-            "Canterbury,\n" + "Kent\n" + "CT2 7NZ\n ", "Campus Watch: +44 (0)1227 868253", "Campus Watch: 07772 226188",
-            "Campus Watch: 07772 226188"};
-}
