+++ app/src/main/java/kent/kentapp/KentUnion.java
@@ -1,0 +1,20 @@
+package kent.kentapp;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.webkit.WebView;
+import android.webkit.WebViewClient;
+
+public class KentUnion extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        //setContentView(R.layout.activity_campus_map);
+        WebView webview = new WebView(this);
+        setContentView(webview);
+        webview.setWebViewClient(new WebViewClient());
+        webview.getSettings().setJavaScriptEnabled(true);
+        webview.loadUrl("http://www.kentunion.co.uk/");
+    }
+}
