+++ app/src/main/java/kent/kentapp/LoginActivity.java
@@ -43,8 +43,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 
-import javax.net.ssl.HttpsURLConnection;
-
 import static android.Manifest.permission.READ_CONTACTS;
 
 /**
@@ -346,10 +344,10 @@ public class LoginActivity extends AppCompatActivity implements LoaderCallbacks<
                 System.out.println("Error fetching user data: " + e);
                 return false;
             }
-            HttpsURLConnection urlConnection;
+            HttpURLConnection urlConnection;
             ArrayList response = new ArrayList();
             try {
-                urlConnection = (HttpsURLConnection) url.openConnection();
+                urlConnection = (HttpURLConnection) url.openConnection();
                 urlConnection.setRequestMethod("GET");
                 //urlConnection.setRequestProperty("username", username);
