+++ app/src/main/res/layout/activity_friend_profile.xml
@@ -68,39 +68,24 @@
             android:layout_row="3"
             android:background="@color/common_action_bar_splitter" />
 
-
         <TextView
-            android:id="@+id/services"
+            android:id="@+id/about_you"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_column="0"
             android:layout_columnSpan="7"
             android:layout_gravity="end"
-            android:layout_marginBottom="13dp"
+            android:layout_marginBottom="5dp"
             android:layout_marginEnd="10dp"
             android:layout_row="4"
-            android:width="230dp"
+            android:width="290dp"
             android:height="24dp"
             android:background="@color/kent_translucent"
             android:ems="10"
-            android:text="@string/services"
+            android:text="@string/person_info"
             android:textAlignment="viewEnd"
             android:textSize="20sp" />
 
-        <EditText
-            android:id="@+id/add_bio"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_column="0"
-            android:layout_columnSpan="7"
-            android:layout_marginBottom="2dp"
-            android:layout_row="6"
-            android:layout_rowSpan="2"
-            android:height="90dp"
-            android:ems="10"
-            android:hint="@string/add_bio"
-            android:textColor="@color/kent_blue"
-            android:textStyle="bold" />
 
         <EditText
             android:id="@+id/add_degree"
@@ -116,10 +101,24 @@
             android:textColor="@color/kent_blue"
             android:textStyle="bold" />
 
+        <EditText
+            android:id="@+id/add_bio"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_column="0"
+            android:layout_columnSpan="7"
+            android:layout_marginBottom="2dp"
+            android:layout_row="6"
+            android:layout_rowSpan="1"
+            android:height="90dp"
+            android:ems="10"
+            android:hint="@string/add_bio"
+            android:textColor="@color/kent_blue"
+            android:textStyle="bold" />
 
 
         <TextView
-            android:id="@+id/about_you"
+            android:id="@+id/friends"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_column="0"
@@ -127,15 +126,99 @@
             android:layout_gravity="end"
             android:layout_marginBottom="5dp"
             android:layout_marginEnd="10dp"
-            android:layout_marginTop="15dp"
             android:layout_row="7"
             android:width="290dp"
             android:height="24dp"
             android:background="@color/kent_translucent"
             android:ems="10"
-            android:text="@string/person_info"
+            android:text="@string/friends"
             android:textAlignment="viewEnd"
             android:textSize="20sp" />
+
+        <ScrollView
+            android:layout_width="match_parent"
+            android:layout_height="280dp"
+            android:layout_column="1"
+            android:layout_columnSpan="7"
+            android:layout_row="8"
+            android:layout_rowSpan="3">
+
+
+            <RelativeLayout
+                android:id="@+id/friends_layout"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:orientation="horizontal">
+
+                <GridLayout
+                    android:id="@+id/gridLayout3"
+                    android:layout_width="wrap_content"
+                    android:layout_height="match_parent"
+                    android:layout_alignParentEnd="true"
+                    android:layout_alignParentStart="true"
+                    android:columnCount="8"
+                    android:rowCount="4">
+
+                    <ImageButton
+                        android:id="@+id/friendBtn"
+                        style="@style/small_user_image"
+                        android:layout_marginStart="30dp"
+                        android:layout_column="1"
+                        android:layout_row="0"
+                        android:contentDescription="@string/friends_recommendations"
+                        android:src="@drawable/avatar2" />
+
+                    <TextView
+                        android:layout_column="1"
+                        android:layout_row="1"
+                        style="@style/small_user_text"
+                        android:layout_marginStart="30dp"/>
+
+                    <ImageButton
+                        android:id="@+id/society_pic2"
+                        style="@style/small_user_image"
+                        android:layout_column="2"
+                        android:layout_row="0"
+                        android:contentDescription="@string/friends_recommendations"
+                        android:src="@drawable/avatar2" />
+
+                    <TextView
+                        android:layout_column="2"
+                        android:layout_row="1"
+                        style="@style/small_user_text" />
+
+                    <ImageButton
+                        android:id="@+id/society_pic3"
+                        style="@style/small_user_image"
+                        android:layout_column="3"
+                        android:layout_row="0"
+                        android:contentDescription="@string/friends_recommendations"
+                        android:src="@drawable/avatar2" />
+
+                    <TextView
+                        android:layout_column="3"
+                        android:layout_row="1"
+                        style="@style/small_user_text"/>
+
+                    <ImageButton
+                        android:id="@+id/society_pic4"
+                        style="@style/small_user_image"
+                        android:layout_column="4"
+                        android:layout_row="0"
+                        android:contentDescription="@string/friends_recommendations"
+                        android:src="@drawable/avatar2" />
+
+                    <TextView
+                        android:layout_column="4"
+                        android:layout_row="1"
+                        style="@style/small_user_text" />
+
+                </GridLayout>
+
+            </RelativeLayout>
+
+        </ScrollView>
+
 
 
 
@@ -165,7 +248,7 @@
                 android:id="@+id/newsBtn"
                 android:layout_width="48dip"
                 android:layout_height="54dip"
-                android:layout_marginStart="17dp"
+                android:layout_marginStart="12dp"
                 android:background="@drawable/news_icon"
                 android:contentDescription="@string/news" />
