+++ app/src/main/res/layout/activity_profile.xml
@@ -102,15 +102,13 @@
 
         <Button
             android:id="@+id/btnMoodle"
-            android:layout_width="78dp"
-            android:layout_height="35dp"
+            android:layout_width="80dp"
+            android:layout_height="40dp"
             android:layout_column="0"
             android:layout_columnSpan="2"
             android:layout_marginBottom="5dp"
-            android:layout_marginStart="26dp"
+            android:layout_marginStart="24dp"
             android:layout_row="5"
-            android:width="29dp"
-            android:height="24dp"
             android:background="@color/kent_blue"
             android:clickable="true"
             android:ems="10"
@@ -122,14 +120,12 @@
         <Button
             android:id="@+id/btnSDS"
             android:layout_width="66dp"
-            android:layout_height="35dp"
+            android:layout_height="40dp"
             android:layout_column="2"
             android:layout_columnSpan="2"
             android:layout_marginBottom="5dp"
             android:layout_marginStart="16dp"
             android:layout_row="5"
-            android:width="29dp"
-            android:height="24dp"
             android:background="@color/kent_blue"
             android:clickable="true"
             android:ems="10"
@@ -141,14 +137,12 @@
         <Button
             android:id="@+id/calendar_button"
             android:layout_width="95dp"
-            android:layout_height="35dp"
+            android:layout_height="40dp"
             android:layout_column="6"
             android:layout_columnSpan="2"
             android:layout_marginBottom="5dp"
             android:layout_marginStart="28dp"
             android:layout_row="5"
-            android:width="29dp"
-            android:height="24dp"
             android:background="@color/kent_blue"
             android:clickable="true"
             android:ems="10"
@@ -230,7 +224,7 @@
                 android:id="@+id/newsBtn"
                 android:layout_width="48dip"
                 android:layout_height="54dip"
-                android:layout_marginStart="17dp"
+                android:layout_marginStart="12dp"
                 android:background="@drawable/news_icon"
                 android:contentDescription="@string/news" />
